Transform read - not-read  into a button.
When you click on the button the status change from read to not read
and vice versa.

we put a button id linked to the index like we did with the delete button.

And we copied the function addEventListenerToButtonX() to add the event
listener to read button -> AddEventListenerToReadButton()
so when we create the html card we put the event listener on the button.

AddEventListenerToReadButton()
  -> set up the listener.
  -> change the boolean Book.read property to the opposite.
  -> Actualize all the cards.

--> only issue, when we actualize the read - not-read status we delete all the
cards and we make them again, it could be more efficient.